Take `&self` instead of `&mut self` in write methods
Taking `&mut self` is not possible because the compiler will make too many guarantees to LLVM in this case, e.g. that spurious reads of the wrapped value are permitted.

Because of this change, the `VolatileCell` type must be `!Sync`. However, it can still be `Send`, so it can still be used in statics when it's wrapped in a `Mutex`.